backend/tstracker/requirements.txt
django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.1.0